\name{UN}
\alias{UN}
\docType{data}
\title{ Unemployment rate, 1890-1979 }
\description{
U.S. unemployment rate
}
\usage{data(UN)}
\format{ ts object}
  
  \details{
Data from 1981. feed: April; gold: June, pork: March, us: March
}

\source{
Data from the package lmtest,  unemployment
}

\references{
J. D. Rea (1983), The explanatory power of alternative theories of inflation and unemployment,
1895-1979. Review of Economics and Statistics 65, 183-195
}


\examples{
plot(UN)
}
\keyword{datasets}
