\name{DiffBA}
\alias{DiffBA}
\docType{data}
\title{
Bond yield differences, annual
}
\description{
The difference in Baa and Aaa corporate bond yields annually
}
\usage{data(DiffBA)}
\format{
  The format is:
 num [1:35] 1.32 0.95 0.76 1.06 1.73 ...
}
\details{
Ying, please provide details

}
\source{
Ying, please provide the data source
}
\references{
Ying, any reference?
}
\examples{
mleurDiag(DiffBA)
mleur(DiffBA)
dftest(DiffBA)
}
\keyword{datasets}
