\name{mleur-package}
\alias{mleur-package}
\docType{package}
\title{Maximum likelihood unit root test}
\description{
Support for mle unit root tests
}

\details{
\tabular{ll}{
Package: \tab mleur\cr
Type: \tab Package\cr
Version: \tab 1.0-3\cr
Date: \tab 2011-08-19\cr
License: GPL (>= 2)\cr
LazyLoad: \tab yes\cr
}
}

\author{
A. I. McLeod, Hao Yu and Ying Zhang

Maintainer: <aimcleod@uwo.ca>
}

\keyword{ package }
\keyword{ ts }

\examples{
mleur(vel)
}
